Finished deleteButtonTapped function and made it so that if the teamName == Chiefs, the color of the cell turns red and if it's Eagles, it will turn teal! Else it will remain white.

+++ Assessment 2/Controllers/Model Controller/TeamController.swift
@@ -14,17 +14,21 @@ class TeamController {
     static let sharedInstance = TeamController()
     var teams: [Team] = []
     
-    func createTeam(name: String, rating: Double, count: Int) {
+    func createTeam(name: String, rating: Int, count: Int) {
         let team = Team(teamName: name.uppercased(), teamRating: rating, teamCount: count)
         teams.append(team)
     }
+    
     // If I want to update a team, I need the new values for the properties AND I need to know WHAT team the user is trying to update...
-    func updateTeam(teamToUpdate: Team, nameToUpdate: String, ratingToUpdate: Double, countToUpdate: Int) {
+    
+    func updateTeam(teamToUpdate: Team, nameToUpdate: String, ratingToUpdate: Int, countToUpdate: Int) {
         teamToUpdate.teamName = nameToUpdate.uppercased()
         teamToUpdate.teamRating = ratingToUpdate
         teamToUpdate.teamCount = countToUpdate
     }
+    
     // If I want to remove a team, I need to know which one it is!
+    
     func deleteTeam(teamToDelete: Team) {
         guard let index = teams.firstIndex(of: teamToDelete) else {return}
         teams.remove(at: index)
